+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Added: `at-mixin-no-argumentless-call-parentheses` rule.
 - Added: `at-import-no-partial-leading-underscore` rule.
 - Added: `at-import-no-partial-extension` rule.
+- Added: `percent-placeholder-pattern` rule.
 
 # 1.0.0